(add) lists and positions details loading spinner and titles
