--- src/pages/PositionDetail/PositionDetailNotFound.tsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export const PositionDetailNotFound = () => {
-  return <h3>Position not found</h3>
-}
